+++ data-raw/sample_data.R
@@ -16,8 +16,10 @@ if(have_auth()) {
   file.remove("./inst/extdata/SG-3115BBBK0782.motus")
   tagme("SG-3115BBBK0782", new = TRUE, update = TRUE, dir = "./inst/extdata/")
   
+  # Update project 4
   #file.remove("./inst/extdata/project-4.motus")
   #tagme(4, new = TRUE, update = TRUE, dir = "./inst/extdata/")
+  tagme(4, new = FALSE, update = TRUE, dir = "./inst/extdata/")
   
   # Create small sample for GPS tests
   file.copy("./inst/extdata/project-4.motus", "./inst/extdata/gps_sample.motus", overwrite = TRUE)