+++ data-raw/updatesql.R
@@ -1,14 +1,52 @@
 sql_versions <- dplyr::tibble()
 
+
 sql_versions <- rbind(
   sql_versions,
-  cbind(date = "2020-04-22",
+  cbind(date = "2020-08-07",
         descr = "Rename nodeDataId to nodeDataID in nodeData",
         sql = paste0("ALTER TABLE nodeData RENAME TO nodeData2;",
                      makeTables(type = "nodeData"),
                      "INSERT INTO nodeData SELECT * FROM nodeData2;",
-                     "DROP TABLE nodeData2;"))
-)
+                     "DROP TABLE nodeData2;")))
+
+sql_versions <- rbind(
+  sql_versions,
+  cbind(date = "2020-08-06", 
+        descr = "Add new CTT V2 fields to 'gps', 'nodeData' and 'hits' tables",
+        sql = paste0("ALTER TABLE gps ADD COLUMN lat_mean FLOAT;",
+                     "ALTER TABLE gps ADD COLUMN lon_mean FLOAT;",
+                     "ALTER TABLE gps ADD COLUMN n_fixes INTEGER;",
+                     "ALTER TABLE nodeData ADD COLUMN nodets FLOAT;",
+                     "ALTER TABLE nodeData ADD COLUMN firmware VARCHAR(20);",
+                     "ALTER TABLE nodeData ADD COLUMN solarVolt FLOAT;",
+                     "ALTER TABLE nodeData ADD COLUMN solarCurrent FLOAT;",
+                     "ALTER TABLE nodeData ADD COLUMN solarCurrentCumul FLOAT;",
+                     "ALTER TABLE nodeData ADD COLUMN lat FLOAT;",
+                     "ALTER TABLE nodeData ADD COLUMN lon FLOAT;",
+                     "ALTER TABLE hits ADD COLUMN validated TINYINT;")))
+
+sql_versions <- rbind(
+  sql_versions,
+  cbind(date = "2020-05-22",
+        descr = "Add 'antFreq' 'antDeps' table",
+        sql = "ALTER TABLE antDeps ADD COLUMN antFreq REAL;"))
+
+sql_versions <- rbind(
+  sql_versions,
+  cbind(date = "2020-04-24",
+        descr = "Add 'test' metadata to 'tagDeps' table",
+        sql = paste0("ALTER TABLE tagDeps ADD COLUMN test INTEGER;",
+                     # Dropped views are recreated in later steps
+                     "DROP VIEW IF EXISTS alltags;",
+                     "DROP VIEW IF EXISTS alltagsGPS;")))
+
+sql_versions <- rbind(
+  sql_versions,
+  cbind(date = "2020-04-23",
+        descr = "Add 'sex' and 'age' metadata to 'tagDeps' table",
+        sql = paste0("ALTER TABLE tagDeps ADD COLUMN sex TEXT;",
+                     "ALTER TABLE tagDeps ADD COLUMN age TEXT;")))
 
 sql_versions <- rbind(
   sql_versions,