+++ tests/testthat/test_10_compare.R
@@ -8,6 +8,14 @@ test_that("table fields match server (sample auth)", {
     system.file("extdata", "project-176.motus", package = "motus"))
 
   expect_named(
+    srvActivityForAll(batchID = 53)[1,],
+    DBI::dbListFields(tags, "activity"), ignore.order = TRUE)
+  
+  #expect_named(
+  #  srvActivityForAll(batchID = 53)[1,],
+  #  DBI::dbListFields(tags, "activityAll"), ignore.order = TRUE)
+  
+  expect_named(
     srvActivityForBatches(batchID = 53)[1,],
     DBI::dbListFields(tags, "activity"), ignore.order = TRUE)
   
@@ -24,13 +32,17 @@ test_that("table fields match server (sample auth)", {
     DBI::dbListFields(tags, "projAmbig"), ignore.order = TRUE)
   
   expect_named(
-    cbind(srvRunsForTagProject(projectID = 176, batchID = 53, runID = 8886)[1,],
-          "nodeNum" = NA),
+    srvRunsForTagProject(projectID = 176, batchID = 53, runID = 8886)[1,],
     DBI::dbListFields(tags, "runs"), ignore.order = TRUE)
   
   expect_named(
-    srvGPSforTagProject(projectID = 176, batchID = 53, gpsID = 0)[1,],
+    srvGPSForTagProject(projectID = 176, batchID = 53, gpsID = 0)[1,],
     DBI::dbListFields(tags, "gps"), ignore.order = TRUE)
+  
+  # Update once sample data has nodeData
+  #expect_named(
+  #  srvNodes(projectID = 176, batchID = 53, nodeDataID = 0)[1,],
+  #  c(DBI::dbListFields(tags, "nodeData"), "projectID"), ignore.order = TRUE)
 })
 
 test_that("table fields match server (local auth)", {
@@ -61,8 +73,12 @@ test_that("table fields match server (local auth)", {
     DBI::dbListFields(tags, "runs"), ignore.order = TRUE)
   
   expect_named(
-    srvGPSforTagProject(projectID = 4, batchID = 53, gpsID = 0)[1,],
+    srvGPSForTagProject(projectID = 4, batchID = 53, gpsID = 0)[1,],
     DBI::dbListFields(tags, "gps"), ignore.order = TRUE)
+  
+  expect_named(
+    srvNodes(projectID = 4, batchID = 53, nodeDataID = 0)[1,],
+    c(DBI::dbListFields(tags, "nodeData"), "projectID"), ignore.order = TRUE)
 })